% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_classes.R
\name{summarise_classes}
\alias{summarise_classes}
\title{Summarise classes}
\usage{
summarise_classes(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe summarising the column classes
}
\description{
Summarise classes
}
